Capybara js emulation handles clicks on links with method post better
This commit fixes the following: Before, the trigger for when the method
click_with_rails_javascript_emulation should intercept clicks was when
the onclick js build a form which had an input element with name set to
_method.

It turned out that this was kinda not working when you had defined a
link like this: link_to(some_path, :method => :post). In these cases
there are no need for Rails to set _method, thus clicking on a link
like this was never intercepted as it should have.